requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
requests==2.31.0
pika==1.3.2
python-dotenv==1.0.0
marshmallow==3.20.1
